Fetch moods (scenes) and merge them into groups.
The groups do not contain moodIDs (like deviceIDs) so I have to add these IDs to the groups myself. The logging is to detect race conditions but so far it seems that the groups are generating events before to moods.
In the client the moods are merged into the groupedDevices computed property